+++ src/dipdup/cli.py
@@ -447,7 +447,7 @@ async def schema_init(ctx):
 
         # NOTE: It's not necessary a reindex, but it's safe to execute built-in scripts to (re)create views.
         conn = get_connection()
-        await generate_schema(conn, config.database.schema_name)
+        await generate_schema(conn, config.database.schema_name, config.advanced.head_status_timeout)
 
     _logger.info('Schema initialized')