+++ docs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [unreleased]
+
+### Added
+
+* context: Added `execute_sql_query` method to run queries from `<project>/sql` directory.
+
+### Fixed
+
+* context: Fixed `execute_sql_scripts` method when the path is a file.
+
+### Deprecated
+
+* context: `execute_sql` is deprecated, use `execute_sql_scripts` instead.
+
+### Performance
+
+* cli: Lazy import dependencies to speed up startup.
+
 ## 5.1.5 - 2022-06-05
 
 ### Fixed